+++ src/workers/metacity/data.ts
@@ -1,0 +1,17 @@
+import { MeshData, PointData } from "../../geometry/data";
+
+export interface InputData {
+    file: string;
+    idOffset: number;
+    styles: string[];
+    baseColor: number;
+}
+
+export interface ParsedData {
+    metadata: {
+        id: number;
+        [id: number]: any;
+    };
+    mesh: MeshData | undefined;
+    points: PointData | undefined;
+}
